+++ src/components/editor/feature-components/MetaformMultiChoiceFieldPropertiesComponent.tsx
@@ -1,6 +1,6 @@
 import { Button, FormControlLabel, IconButton, Stack, Switch, TextField, Tooltip, Typography } from "@mui/material";
 import { uuid4 } from "@sentry/utils";
-import { FieldRule, Metaform, MetaformField, MetaformFieldOption } from "generated/client";
+import { MetaformField, MetaformFieldOption } from "generated/client";
 import produce from "immer";
 import strings from "localization/strings";
 import React, { FC, useState } from "react";
@@ -14,103 +14,21 @@ import { useAppSelector, useAppDispatch } from "app/hooks";
  * Component properties
  */
 interface Props {
-  setPendingForm: (metaform: Metaform) => void;
-  setUpdatedMetaformField: (updatedMetaformField: MetaformField) => void;
+  updateFormFieldDebounced: (updatedField: MetaformField, optionIndex?: number) => void;
 }
 
 /**
  * Draft editor right drawer feature define member group component
  */
 const MetaformMultiChoiceFieldComponent: FC<Props> = ({
-  setPendingForm,
-  setUpdatedMetaformField
+  updateFormFieldDebounced
 }) => {
   const [ multiSelectRawTextMode, setMultiSelectRawTextMode ] = useState<boolean>(false);
   const [ multiSelectRawText, setMultiSelectRawText ] = useState<string>("");
   const [ memberGroupOptIndex, setMemberGroupOptIndex ] = useState<number>();
-  const [ debounceTimerId, setDebounceTimerId ] = useState<NodeJS.Timeout>();
   const { metaformSectionIndex, metaformFieldIndex, metaformVersion, metaformField } = useAppSelector(selectMetaform);
   const pendingForm = MetaformUtils.jsonToMetaform(MetaformUtils.getDraftForm(metaformVersion));
   const dispatch = useAppDispatch();
-
-  /**
-   * 
-   * @param fieldRules field rules
-   * @param field metaform field
-   * @param optionIndex option index
-   * @param fieldOptionMatch field option match
-   */
-  const checkFieldRules = (fieldRules: FieldRule[], field: MetaformField, optionIndex?: number, fieldOptionMatch?: MetaformFieldOption) => {
-    fieldRules.forEach(rule => {
-      if ((metaformField!.name !== undefined && field.name !== metaformField!.name)) {
-        rule.field = field.name;
-      // option update
-      } else if (optionIndex !== undefined) {
-        const fieldOptionToUpdate = field.options![optionIndex];
-        if (rule.equals === fieldOptionMatch!.name) {
-          rule.equals = fieldOptionToUpdate.name;
-        } else if (rule.notEquals === fieldOptionMatch!.name) {
-          rule.notEquals = fieldOptionToUpdate.name;
-        }
-      }
-      return field;
-    });
-  };
-
-  /**
-   * Updates field with visibility
-   *
-   * @param field edited field
-   * @param optionIndex option index
-   */
-  const updateFormField = (field: MetaformField, optionIndex?: number) => {
-    if (!metaformField) {
-      return;
-    }
-
-    const updatedForm = produce(pendingForm, draftForm => {
-      if (MetaformUtils.fieldTypesAllowVisibility.includes(field.type)) {
-        if ((metaformField.name !== undefined && field.name !== metaformField.name) || optionIndex !== undefined) {
-          const fieldOptionMatch = optionIndex !== undefined ?
-            pendingForm.sections![metaformSectionIndex!].fields![metaformFieldIndex!].options![optionIndex] :
-            undefined;
-          const fieldNameMatch = pendingForm.sections![metaformSectionIndex!].fields![metaformFieldIndex!].name || "";
-
-          const fieldRules: FieldRule[] = [];
-
-          draftForm.sections?.forEach(draftSection => {
-            if (draftSection.visibleIf !== undefined) {
-              MetaformUtils.fieldRuleScan(draftSection.visibleIf, fieldNameMatch, fieldRules, fieldOptionMatch);
-            }
-            draftSection.fields?.forEach(draftField => {
-              if (draftField.visibleIf !== undefined) {
-                MetaformUtils.fieldRuleScan(draftField.visibleIf, fieldNameMatch, fieldRules, fieldOptionMatch);
-              }
-            });
-          });
-          checkFieldRules(fieldRules, field, optionIndex, fieldOptionMatch);
-        }
-      }
-
-      draftForm.sections?.[metaformSectionIndex!]?.fields?.splice(metaformFieldIndex!, 1, field);
-    });
-
-    setPendingForm(updatedForm);
-  };
-
-  /**
-   * Debounced update field
-   *
-   * @param field edited field
-   * @param optionIndex option index
-   */
-  const updateFormFieldDebounced = (field: MetaformField, optionIndex?: number) => {
-    dispatch(setMetaformField(field));
-
-    debounceTimerId && clearTimeout(debounceTimerId);
-    setDebounceTimerId(setTimeout(() => updateFormField(field, optionIndex), 500));
-  };
-
   /**
    * Update option text and check if its used as VisibleIf change text also in that field
    *
@@ -126,6 +44,7 @@ const MetaformMultiChoiceFieldComponent: FC<Props> = ({
       draftField?.options?.splice(optionIndex, 1, updateTextOption);
     });
 
+    dispatch(setMetaformField(updatedField));
     updateFormFieldDebounced(updatedField, optionIndex);
   };
 
@@ -148,28 +67,9 @@ const MetaformMultiChoiceFieldComponent: FC<Props> = ({
     const updatedField: MetaformField = produce(metaformField, draftField => {
       draftField?.options?.splice(optionIndex, 1);
     });
+
+    dispatch(setMetaformField(updatedField));
     updateFormFieldDebounced(updatedField);
-
-    const updatedForm = produce(pendingForm, draftForm => {
-      draftForm.sections?.forEach(draftSection => {
-        if (draftSection.visibleIf &&
-          MetaformUtils.fieldRuleMatch(draftSection.visibleIf, fieldNameMatch, optionMatch)
-        ) {
-          draftSection.visibleIf = undefined;
-        }
-
-        draftSection.fields?.forEach(draftField => {
-          if (draftField.visibleIf &&
-            MetaformUtils.fieldRuleMatch(draftField.visibleIf, fieldNameMatch, optionMatch)
-          ) {
-            draftField.visibleIf = undefined;
-          }
-        });
-      });
-
-      draftForm.sections?.[metaformSectionIndex]?.fields?.[metaformFieldIndex]?.options?.splice(optionIndex, 1);
-    });
-    setPendingForm(updatedForm);
   };
 
   /**
@@ -225,6 +125,7 @@ const MetaformMultiChoiceFieldComponent: FC<Props> = ({
       draftField.options = [ ...(draftField.options || []), newOption ];
     });
 
+    dispatch(setMetaformField(updatedField));
     updateFormFieldDebounced(updatedField);
   };
 
@@ -265,7 +166,7 @@ const MetaformMultiChoiceFieldComponent: FC<Props> = ({
       }));
     });
     dispatch(setMetaformField(updatedField));
-    setUpdatedMetaformField(updatedField);
+    updateFormFieldDebounced(updatedField);
     setMultiSelectRawTextMode(false);
   };
